# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Music Player API
  description: API for managing music playlists (groups), tracks, and playback board
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  parameters:
...
      UserId:
        name: userId
        in: path
        description: ID of the user
        required: true
        schema:
          type: integer
          format: int64
          example: 1
      BoardId:
        name: boardId
        in: path
        description: ID of the board
        required: true
        schema:
          type: integer
          format: int64
          example: 1
      GroupId:
        name: groupId
        in: path
        description: ID of the group
        required: true
        schema:
          type: integer
          format: int64
          example: 1
  schemas:
    User:
      type: object
      description: Basic user representation
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the user
          example: 1
        name:
          type: string
          description: Unique username
          example: "bruno"
...
    GroupShareRequest:
      type: object
      description: Request to share a group with another user
      required:
        - userId
      properties:
        userId:
          type: integer
          format: int64
          description: ID of the user to share the group with
          example: 2
    GroupShare:
      type: object
      description: Group sharing information for a specific user
      properties:
        user:
          $ref: '#/components/schemas/User'
    Track:
      type: object
      description: A music track
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier (auto-generated)
          example: 1
        trackName:
          type: string
          description: Name or title of the track
          example: "Sample Song"
        trackLink:
          type: string
          description: URL or link to the track audio
          example: "https://example.com/track.mp3"
        duration:
          type: integer
          description: Duration of the track in seconds
          example: 215
        ownerId:
          type: integer
          format: int64
          description: ID of the user who owns this track
          example: 1
        groupIds:
          type: array
          description: IDs of groups this track belongs to
          items:
            type: integer
            format: int64
          example: [3, 7, 12]
        trackPoints:
          type: array
          description: Points of interest for the track
          items:
            $ref: '#/components/schemas/TrackPoint'
    TrackPoint:
      type: object
      description: Point of interest in track
      properties:
          id:
            type: integer
            format: int64
            description: Unique identifier (auto-generated)
          name:
              type: string
              description: name for point
          position:
            type: integer
            description: Point of interest in the track
          fadeIn:
            type: boolean
            description: point makes a fade in
          fadeOut:
            type: boolean
            description: point makes a fade out
...
    Group:
      type: object
      description: A playlist or group of tracks
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier (auto-generated)
          example: 1
        listName:
          type: string
          description: Name of the playlist or group
          example: "My Favorites"
        ownerId:
          type: integer
          format: int64
          description: ID of the user who owns this group
          example: 1
        tracks:
          type: array
          description: Tracks in this group
          items:
            $ref: '#/components/schemas/Track'
    GroupRequest:
      type: object
      description: Request to create or update a group
      required:
        - listName
      properties:
        listName:
          type: string
          description: Name of the playlist or group
          example: "My Favorites"
        ownerId:
          type: integer
          format: int64
          description: owner id of group
          example: 1
        tracks:
          type: array
          description: Tracks in this group
          items:
            $ref: '#/components/schemas/Track'
    Board:
      type: object
      description: A playback board
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier (auto-generated)
          readOnly: true
          example: 1
        userId:
          type: integer
          format: int64
          description: ID of the owning user
          readOnly: true
          example: 1
        name:
          type: string
          description: Name of the board
          example: "Default Playback"
        selectedTrack:
          $ref: '#/components/schemas/Track'
        volume:
          type: integer
          description: Playback volume (0-100)
          minimum: 0
          maximum: 100
          example: 50
        repeat:
          type: boolean
          description: Repeat playback mode
          example: false
        overplay:
          type: boolean
          description: Overplay/overlay mode
          example: false
        ownerId:
          type: integer
          format: int64
          description: owner id of group
          example: 1
    BoardCreateRequest:
      type: object
      description: Request to create a new board
      required:
        []
      properties:
        name:
          type: string
          description: Name of the board
          example: "Default Playback"
        selectedTrack:
          $ref: '#/components/schemas/Track'
        volume:
          type: integer
          description: Playback volume (0-100)
          minimum: 0
          maximum: 100
          example: 50
        repeat:
          type: boolean
          description: Repeat playback mode
          example: false
        overplay:
          type: boolean
          description: Overplay/overlay mode
          example: false
    BoardUpdateRequest:
      type: object
      description: Update for a board
      required: []
      properties:
        selectedTrack:
          $ref: '#/components/schemas/Track'
        volume:
          type: integer
          description: Playback volume (0-100)
          minimum: 0
          maximum: 100
          example: 50
        name:
          type: string
          description: Name of the board
          example: "Default Playback"
        repeat:
          type: boolean
          description: Repeat playback mode
          example: false
        overplay:
          type: boolean
          description: Overplay/overlay mode
          example: false
    PlayRequest:
      type: object
      description: >
        Request to start playback on a board. If trackId is omitted, plays the board's selectedTrack.
      properties:
        trackId:
          type: integer
          format: int64
          nullable: true
          description: Track to play on the board (optional)
          example: 1
    SeekRequest:
      type: object
      description: Request to seek within the currently playing track
      required:
        - positionS
      properties:
        positionS:
          type: integer
          format: int64
          minimum: 0
          description: Target position in seconds from start of track
          example: 60
...
    PlaybackStatus:
      type: string
      description: Current playback state of the board
      enum:
        - STOPPED
        - PLAYING
        - PAUSED
        - BUFFERING
        - ERROR
    PlaybackState:
      type: object
      description: Runtime playback state for a board
      required:
        - boardId
        - status
      properties:
        boardId:
          type: integer
          format: int64
          example: 1
        status:
          $ref: '#/components/schemas/PlaybackStatus'
        trackId:
          type: integer
          format: int64
          nullable: true
          description: Currently loaded/playing track id
          example: 1
        positionS:
          type: integer
          format: int64
          nullable: true
          description: Current playback position in seconds (if known)
          example: 12
        windowStartS:
          type: integer
          format: int64
          nullable: true
          description: Active window start in seconds (if set)
          example: 30
        windowEndS:
          type: integer
          format: int64
          nullable: true
          description: Active window end in seconds (if set)
          example: 90
        volume:
          type: integer
          nullable: true
          minimum: 0
          maximum: 100
          example: 50
        repeat:
          type: boolean
          nullable: true
          example: false
        overplay:
          type: boolean
          nullable: true
          example: false
paths:
...
  /boards:
    get:
      tags:
        - MusicBoards
      summary: Get all boards for a user
      operationId: getUserBoards
      responses:
        '200':
          description: Successful response with list of boards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
        '404':
          description: User not found
    post:
      tags:
        - MusicBoards
      summary: Create a new board for a user
      operationId: createUserBoard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardCreateRequest'
      responses:
        '201':
          description: Board created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '400':
          description: Invalid input
        '404':
          description: User not found
  /boards/{boardId}:
    get:
      tags:
        - MusicBoards
      summary: Get a specific board
      operationId: getUserBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '200':
          description: Successful response with board details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          description: Board or user not found
    put:
      tags:
        - MusicBoards
      summary: Update a board (partial updates supported)
      operationId: updateUserBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardUpdateRequest'
      responses:
        '200':
          description: Board updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '400':
          description: Invalid input
        '404':
          description: Board or user not found
    delete:
      tags:
        - MusicBoards
      summary: Delete a board
      operationId: deleteUserBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '204':
          description: Board deleted successfully
        '404':
          description: Board or user not found
  /groups:
    get:
      tags:
        - MusicGroups
      summary: Get all groups for a user
      operationId: getUserGroups
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '404':
          description: User not found
    post:
      tags:
        - MusicGroups
      summary: Create a new group
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
  /groups/{groupId}:
    delete:
      tags:
        - MusicGroups
      summary: Delete a group
      operationId: deleteGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Group deleted
        '404':
          description: Group not found
    put:
      tags:
        - MusicGroups
      summary: Update a group
      operationId: updateGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRequest'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: Group not found
  /groups/{groupId}/shares:
    get:
      tags:
        - MusicGroups
      summary: List users the group is shared with
      description: Returns all users that have view access to this group. Only the owner can call this endpoint.
      operationId: listGroupShares
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: List of users the group is shared with
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupShare'
        '403':
          description: Forbidden (only the owner can view shares)
        '404':
          description: Group not found
    post:
      tags:
        - MusicGroups
      summary: Share a group with a user
      description: >
        Shares the group with another user. If already shared, this is a no-op.
        Only the owner can call this endpoint.
      operationId: shareGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupShareRequest'
      responses:
        '200':
          description: Group sharing created (or already existed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupShare'
        '400':
          description: Invalid input
        '403':
          description: Forbidden
        '404':
          description: Group or user not found
  /groups/{groupId}/shares/{userId}:
    delete:
      tags:
        - MusicGroups
      summary: Remove sharing of a group for a user
      description: >
        Removes access to the group for the specified user.
      operationId: unshareGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Sharing removed successfully
        '403':
          description: Forbidden
        '404':
          description: Group or sharing not found
  /boards/{boardId}/playback/play:
    post:
      tags:
        - Playback
      summary: Start playback on a board
      description: >
        Starts playback on the given board. If trackId is omitted, plays the board's selectedTrack.
        The board must belong to the authenticated user.
      operationId: playBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayRequest'
      responses:
        '200':
          description: Playback started (or restarted) successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackState'
        '400':
          description: Invalid input
        '401':
          description: User not authenticated
        '403':
          description: Forbidden (board not owned by user or track not accessible)
        '404':
          description: Board or track not found
        '409':
          description: No track selected on board (and no trackId provided)
  /boards/{boardId}/playback/stop:
    post:
      tags:
        - Playback
      summary: Stop playback on a board
      operationId: stopBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '200':
          description: Playback stopped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackState'
        '401':
          description: User not authenticated
        '403':
          description: Forbidden (board not owned by user)
        '404':
          description: Board not found
        '409':
          description: Action not valid in current playback state
  /boards/{boardId}/playback/pause:
    post:
      tags:
        - Playback
      summary: Pause playback on a board
      operationId: pauseBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '200':
          description: Playback paused successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackState'
        '401':
          description: User not authenticated
        '403':
          description: Forbidden (board not owned by user)
        '404':
          description: Board not found
        '409':
          description: Action not valid in current playback state
  /boards/{boardId}/playback/resume:
    post:
      tags:
        - Playback
      summary: Resume playback on a board
      operationId: resumeBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '200':
          description: Playback resumed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackState'
        '401':
          description: User not authenticated
        '403':
          description: Forbidden (board not owned by user)
        '404':
          description: Board not found
        '409':
          description: Action not valid in current playback state
  /boards/{boardId}/playback/seek:
    post:
      tags:
        - Playback
      summary: Seek within the current track on a board
      description: Seeks to the given position in seconds from start of track.
      operationId: seekBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeekRequest'
      responses:
        '200':
          description: Seek applied successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackState'
        '400':
          description: Invalid input (negative position, etc.)
        '401':
          description: User not authenticated
        '403':
          description: Forbidden (board not owned by user)
        '404':
          description: Board not found
        '409':
          description: Action not valid in current playback state
  /boards/{boardId}/playback/state:
    get:
      tags:
        - Playback
      summary: Get current playback state of a board
      operationId: getBoardPlaybackState
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '200':
          description: Current playback state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackState'
        '401':
          description: User not authenticated
        '403':
          description: Forbidden (board not owned by user)
        '404':
          description: Board not found
  /boards/{boardId}/stream:
    get:
      tags:
        - Playback
      summary: Stream audio for a board
      description: >
        Streams the currently playing board audio as an MP3 stream.
        Requires authentication. Board must belong to the authenticated user.
      operationId: streamBoardAudio
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '200':
          description: MP3 audio stream
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '401':
          description: User not authenticated
        '403':
          description: Forbidden (board not owned by user)
        '404':
          description: Board not found
        '409':
          description: Board is not playing